# service/api/api.yaml
openapi: 3.0.3
info:
  title: Todo Service
  description: |-
    This is a sample todo service 
  version: 1.0.0
paths:
  /tasks:
    get:
      operationId: getTasks
      description: Returns a list of tasks
      parameters:
        - name: page
          in: query
          description: Zero-based page index
          schema:
            type: integer
        - name: size
          in: query
          description: The size of the page to return
          schema:
            type: integer
      responses:
        '200':
          description: Successfully returned a list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      operationId: createTask
      description: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                username:
                  type: string
      responses:
        '200':
          description: Successfully created a new task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  /tasks/{task_id}:
    get:
      operationId: getTaskById
      description: Get task by id
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully returned a list of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

components:
  schemas:
    Task:
      type: object
      required:
        - task_id
        - name
      properties:
        task_id:
          type: string
        name:
          type: string
        description:
          type: string
        username:
          type: string
        creation_time:
          type: string
          format: date-time